Schema Table and proposed test
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@5273 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ Crud/SchemaTable.cs
@@ -32,7 +32,7 @@ namespace Crud {
 //		[CrudColumn(CrudSpecialColType=CrudSpecialColType.TimeSpanNeg)]
 		public TimeSpan TimeSpanTest;
 		///<summary>Column of data type byte</summary>
-		public long CurrencyTest;
+		public double CurrencyTest;
 		///<summary>Column of data type byte</summary>
 		public bool BoolTest;
 		///<summary>Column of data type byte</summary>
